Make set/isolate work in BOTH directions
Since the beginning I have had set/isolate be unidirection, so that I could
see the far node's config, but he could not see mine. This is too confusing
for yer modern sysop and I have now made it such that only the local configs
are remembered, regardless of what is sent back to me. I.e. it is now
symmetric and bidirectional.