allow PC92(A/D) to be slugged and aggregated
Wait for a period of time before issuing pc92a or d records
and then aggregate them together as a collection of adds or deletes
in just (up to) two sentences, instead of one per event.

This has the side effect of (nearly) eliminating the "login, fire a
spot, logoff" merchants.