Improve short callsign handling (eg K7A)
The format of cty.dat has changed in that exact calls now have
an = prefixed to them. This allow reliable checking for exact short
calls that are otherwise treated as prefixes.

Also fix a small buglet on console.pl that meant that line went awry on
lines that are exactly the column width of the screen.