Add a ring buffer to debugging + other things
Add a ring buffer to debugging to allow sysops to reduce the amount of pointless debugging info that is
continually logged. Likely this is something that big sites or RPi users might want.

The biggy is 'chan'. You want this in the ring buffer, but not logging. To do this:

set/debug chan nologchan

You can log the ring contents at any time with: show/debug_ring
NOTE: you should probably have watchdbg running for this to be much use unless you are doing this as a
matter of record.

export_users has been fixed and is forced to be dumped into /spider/local_data as default.

show/debug now sorts its output.

LogDbg has been fixed to log stuff to the correct category.