modify RBN timing arrangements
There are two new twiddle pots:

$limbotime (default 5*60 secs) which allows potential spots to
hang around longer to allow $minqual *different* skimmers to spot
them. At which point they are emitted. This is to catch the "slow
burning" spots where a call is spotted by different skimmers but
separated in time by more than $dwelltime - in fact maybe by
minutes.

$maxqual (default 9 skimmers) which short circuits the normal
hard maximum $quality (9 skimmer spots) and $dwelltime (10secs)
to allow a spot that has $maxqual *different* skimmers (usually
as the result of a burst from the RBN) to defeat $dwelltime and
be emitted the moment that condition is satisfied. There maybe
a better name for this.