don't fork sh/dx on simple query
If the query does not contain an "expression" (or even just a
callsign and is garanteed to be satified by the cache - if it is
enabled - then do not fork / run as a sub command.

So 'sh/dx', 'sh/mydx' with or without a count will be run from
the cache - if it is enabled.